+++ examples/bottle_example.py
@@ -19,7 +19,7 @@ from webargs import fields, validate
 from webargs.bottleparser import use_args, use_kwargs
 
 
-hello_args = {"name": fields.Str(missing="Friend")}
+hello_args = {"name": fields.Str(load_default="Friend")}
 
 
 @route("/", method="GET", apply=use_args(hello_args))
@@ -40,7 +40,9 @@ def add(x, y):
 dateadd_args = {
     "value": fields.Date(required=False),
     "addend": fields.Int(required=True, validate=validate.Range(min=1)),
-    "unit": fields.Str(missing="days", validate=validate.OneOf(["minutes", "days"])),
+    "unit": fields.Str(
+        load_default="days", validate=validate.OneOf(["minutes", "days"])
+    ),
 }