+++ examples/tornado_example.py
@@ -37,7 +37,7 @@ class BaseRequestHandler(RequestHandler):
 class HelloHandler(BaseRequestHandler):
     """A welcome page."""
 
-    hello_args = {"name": fields.Str(missing="Friend")}
+    hello_args = {"name": fields.Str(load_default="Friend")}
 
     @use_args(hello_args)
     def get(self, args):
@@ -62,7 +62,7 @@ class DateAddHandler(BaseRequestHandler):
         "value": fields.Date(required=False),
         "addend": fields.Int(required=True, validate=validate.Range(min=1)),
         "unit": fields.Str(
-            missing="days", validate=validate.OneOf(["minutes", "days"])
+            load_default="days", validate=validate.OneOf(["minutes", "days"])
         ),
     }